pyproject.toml
[project]
name = "winidjango"
version = "2.0.75"
description = "A utility package for django"
authors = [
    {name = "Winipedia"},
]
license = "MIT"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django",
    "django-stubs-ext",
    "winiutils",
]
classifiers = [
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Operating System :: OS Independent",
    "Typing :: Typed",
]
license-files = [
    "LICENSE",
]
maintainers = [
    {name = "Winipedia"},
]
keywords = []

[project.scripts]
winidjango = "pyrig.rig.cli.cli:main"

[project.urls]
Homepage = "https://github.com/Winipedia/winidjango"
Documentation = "https://Winipedia.github.io/winidjango"
Source = "https://github.com/Winipedia/winidjango"
Issues = "https://github.com/Winipedia/winidjango/issues"
Changelog = "https://github.com/Winipedia/winidjango/releases"

[dependency-groups]
dev = [
    "bandit>=1.9.3",
    "django-stubs>=5.2.9",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.7.1",
    "mkdocs-mermaid2-plugin>=1.2.3",
    "mkdocstrings[python]>=1.0.3",
    "pip-audit>=2.10.0",
    "prek>=0.3.2",
    "pygithub>=2.8.1",
    "pyrig-dev>=0.1.19",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.11.1",
    "pytest-mock>=3.15.1",
    "ruff>=0.15.0",
    "rumdl>=0.1.18",
    "ty>=0.0.16",
    "types-defusedxml>=0.7.0.20250822",
    "types-tqdm>=4.67.3.20260205",
]

[build-system]
requires = [
    "uv_build",
]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "winidjango"
module-root = ""

[tool.ruff]
exclude = [
    ".*",
    "**/migrations/*.py",
]

[tool.ruff.lint]
select = [
    "ALL",
]
ignore = [
    "D203",
    "D213",
    "COM812",
    "ANN401",
]
fixable = [
    "ALL",
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101",
]
"**/tests/**/*.py" = [
    "S101",
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
strict = true
warn_unreachable = true
show_error_codes = true
files = "."

[tool.pytest.ini_options]
testpaths = [
    "tests",
]
addopts = "--cov=winidjango --cov-report=term-missing --cov-fail-under=90"

[tool.bandit]
exclude_dirs = [
    ".*",
]

[tool.bandit.assert_used]
skips = [
    "*/tests/*.py",
    "*test_*.py",
]

[tool.ty.terminal]
error-on-warning = true

[tool.rumdl]
respect_gitignore = true
